+++ app/controllers/builder/answer_options_controller.rb
@@ -1,18 +1,13 @@
+# frozen_string_literal: true
+
+# Allows survey builders to add answer options to answer templates
 class Builder::AnswerOptionsController < Builder::BuilderController
   before_action :authenticate_user!
   before_action :redirect_non_builders
-
-  before_action :set_editable_survey
-  before_action :redirect_without_survey
-
-  before_action :set_editable_question
-  before_action :redirect_without_question
-
-  before_action :set_editable_answer_template
-  before_action :redirect_without_answer_template
-
-  before_action :set_editable_answer_option,        only: [:show, :edit, :update, :destroy]
-  before_action :redirect_without_answer_option,    only: [:show, :edit, :update, :destroy]
+  before_action :find_editable_survey_or_redirect
+  before_action :find_editable_question_or_redirect
+  before_action :find_editable_answer_template_or_redirect
+  before_action :find_editable_answer_option_or_redirect, only: [:show, :edit, :update, :destroy]
 
   def index
     redirect_to builder_survey_question_answer_template_path(@survey, @question, @answer_template)
@@ -24,16 +19,11 @@ class Builder::AnswerOptionsController < Builder::BuilderController
 
   def create
     @answer_option = @answer_template.answer_options.where(user_id: current_user.id).new(answer_option_params)
-
-    respond_to do |format|
-      if @answer_option.save
-        @answer_template.answer_options << @answer_option
-        format.html { redirect_to builder_survey_question_answer_template_answer_option_path(@survey, @question, @answer_template, @answer_option), notice: 'Answer Option was successfully created.' }
-        format.json { render :show, status: :created, location: @answer_option }
-      else
-        format.html { render :new }
-        format.json { render json: @answer_option.errors, status: :unprocessable_entity }
-      end
+    if @answer_option.save
+      @answer_template.answer_options << @answer_option
+      redirect_to builder_survey_question_answer_template_answer_option_path(@survey, @question, @answer_template, @answer_option), notice: 'Answer Option was successfully created.'
+    else
+      render :new
     end
   end
 
@@ -44,33 +34,34 @@ class Builder::AnswerOptionsController < Builder::BuilderController
   end
 
   def update
-    respond_to do |format|
-      if @answer_option.update(answer_option_params)
-        format.html { redirect_to builder_survey_question_answer_template_answer_option_path(@survey, @question, @answer_template, @answer_option), notice: 'Answer Option was successfully updated.' }
-        format.json { render :show, status: :ok, location: @answer_option }
-      else
-        format.html { render :edit }
-        format.json { render json: @answer_option.errors, status: :unprocessable_entity }
-      end
+    if @answer_option.update(answer_option_params)
+      redirect_to builder_survey_question_answer_template_answer_option_path(@survey, @question, @answer_template, @answer_option), notice: 'Answer Option was successfully updated.'
+    else
+      render :edit
     end
   end
 
   def destroy
     @answer_option.destroy
-    respond_to do |format|
-      format.html { redirect_to builder_survey_question_answer_template_path(@survey, @question, @answer_template), notice: 'Answer Option was successfully destroyed.' }
-      format.json { head :no_content }
+    redirect_to builder_survey_question_answer_template_path(@survey, @question, @answer_template), notice: 'Answer Option was successfully destroyed.'
+  end
+
+  # POST /answer_options/reorder.js
+  def reorder
+    params[:answer_option_ids].each_with_index do |answer_option_id, index|
+      aoat = @answer_template.answer_options_answer_templates.find_by answer_option_id: answer_option_id
+      aoat.update position: index if aoat
     end
+    head :ok
   end
 
   private
 
-    def set_editable_answer_option
-      super(:id)
-    end
+  def find_editable_answer_option_or_redirect
+    super(:id)
+  end
 
-    def answer_option_params
-      params.require(:answer_option).permit(:text, :hotkey, :value, :display_class)
-    end
-
+  def answer_option_params
+    params.require(:answer_option).permit(:text, :hotkey, :value, :display_class)
+  end
 end