+++ app/controllers/builder/survey_encounters_controller.rb
@@ -1,12 +1,11 @@
+# frozen_string_literal: true
+
+# Allows survey builders to add encounters to surveys
 class Builder::SurveyEncountersController <  Builder::BuilderController
   before_action :authenticate_user!
   before_action :redirect_non_builders
-
-  before_action :set_editable_survey
-  before_action :redirect_without_survey
-
-  before_action :set_editable_survey_encounter,     only: [:show, :edit, :update, :destroy]
-  before_action :redirect_without_survey_encounter, only: [:show, :edit, :update, :destroy]
+  before_action :find_editable_survey_or_redirect
+  before_action :find_editable_survey_encounter_or_redirect, only: [:show, :edit, :update, :destroy]
 
   def index
     redirect_to builder_survey_path(@survey)
@@ -18,15 +17,10 @@ class Builder::SurveyEncountersController <  Builder::BuilderController
 
   def create
     @survey_encounter = @survey.survey_encounters.where(user_id: current_user.id).new(survey_encounter_params)
-
-    respond_to do |format|
-      if @survey_encounter.save
-        format.html { redirect_to builder_survey_survey_encounter_path(@survey, @survey_encounter), notice: 'Survey Encounter was successfully created.' }
-        format.json { render :show, status: :created, location: @survey_encounter }
-      else
-        format.html { render :new }
-        format.json { render json: @survey_encounter.errors, status: :unprocessable_entity }
-      end
+    if @survey_encounter.save
+      redirect_to builder_survey_survey_encounter_path(@survey, @survey_encounter), notice: 'Survey Encounter was successfully created.'
+    else
+      render :new
     end
   end
 
@@ -37,37 +31,30 @@ class Builder::SurveyEncountersController <  Builder::BuilderController
   end
 
   def update
-    respond_to do |format|
-      if @survey_encounter.update(survey_encounter_params)
-        format.html { redirect_to builder_survey_survey_encounter_path(@survey, @survey_encounter), notice: 'Survey Encounter was successfully updated.' }
-        format.json { render :show, status: :ok, location: @survey_encounter }
-      else
-        format.html { render :edit }
-        format.json { render json: @survey_encounter.errors, status: :unprocessable_entity }
-      end
+    if @survey_encounter.update(survey_encounter_params)
+      redirect_to builder_survey_survey_encounter_path(@survey, @survey_encounter), notice: 'Survey Encounter was successfully updated.'
+    else
+      render :edit
     end
   end
 
   def destroy
     @survey_encounter.destroy
-    respond_to do |format|
-      format.html { redirect_to builder_survey_path(@survey), notice: 'Survey Encounter was successfully destroyed.' }
-      format.json { head :no_content }
-    end
+    redirect_to builder_survey_path(@survey), notice: 'Survey Encounter was successfully destroyed.'
   end
 
   private
 
-    def set_editable_survey_encounter
-      @survey_encounter = @survey.survey_encounters.find_by_id(params[:id])
-    end
+  def find_editable_survey_encounter_or_redirect
+    @survey_encounter = @survey.survey_encounters.find_by_id(params[:id])
+    redirect_without_survey_encounter
+  end
 
-    def redirect_without_survey_encounter
-      empty_response_or_root_path(builder_survey_path(@survey)) unless @survey_encounter
-    end
+  def redirect_without_survey_encounter
+    empty_response_or_root_path(builder_survey_path(@survey)) unless @survey_encounter
+  end
 
-    def survey_encounter_params
-      params.require(:survey_encounter).permit(:encounter_id, :parent_survey_encounter_id)
-    end
-
+  def survey_encounter_params
+    params.require(:survey_encounter).permit(:encounter_id, :parent_survey_encounter_id)
+  end
 end