+++ app/models/surveys/answer_templates_question.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+# Associates questions and answer templates
+class AnswerTemplatesQuestion < ActiveRecord::Base
+  # Associations
+  belongs_to :question, -> { current }
+  belongs_to :answer_template, -> { current }
+end