+++ db/migrate/20160420154349_add_position_to_answer_option_answer_templates.rb
@@ -1,0 +1,6 @@
+class AddPositionToAnswerOptionAnswerTemplates < ActiveRecord::Migration
+  def change
+    add_column :answr_options_answer_templates, :position, :integer, null: false, default: 0
+    add_index :answr_options_answer_templates, :position
+  end
+end