+++ db/migrate/20160420163338_add_position_to_answer_templates_questions.rb
@@ -1,0 +1,6 @@
+class AddPositionToAnswerTemplatesQuestions < ActiveRecord::Migration
+  def change
+    add_column :answer_templates_questions, :position, :integer, null: false, default: 0
+    add_index :answer_templates_questions, :position
+  end
+end