+++ test/controllers/builder/answer_options_controller_test.rb
@@ -10,7 +10,7 @@ class Builder::AnswerOptionsControllerTest < ActionController::TestCase
     @answer_option = answer_options(:web_answer_template_01)
   end
 
-  test "should get answer options as builder" do
+  test 'should get answer options as builder' do
     login(@builder)
     get :index, survey_id: @survey, question_id: @question, answer_template_id: @answer_template
     assert_not_nil assigns(:survey)
@@ -19,7 +19,7 @@ class Builder::AnswerOptionsControllerTest < ActionController::TestCase
     assert_redirected_to builder_survey_question_answer_template_path(assigns(:survey), assigns(:question), assigns(:answer_template))
   end
 
-  test "should not get answer options as regular user" do
+  test 'should not get answer options as regular user' do
     login(@regular_user)
     get :index, survey_id: @survey, question_id: @question, answer_template_id: @answer_template
     assert_nil assigns(:survey)
@@ -28,7 +28,7 @@ class Builder::AnswerOptionsControllerTest < ActionController::TestCase
     assert_redirected_to root_path
   end
 
-  test "should get new answer option as builder" do
+  test 'should get new answer option as builder' do
     login(@builder)
     get :new, survey_id: @survey, question_id: @question, answer_template_id: @answer_template
     assert_not_nil assigns(:survey)
@@ -38,7 +38,7 @@ class Builder::AnswerOptionsControllerTest < ActionController::TestCase
     assert_response :success
   end
 
-  test "should not get new answer option as regular user" do
+  test 'should not get new answer option as regular user' do
     login(@regular_user)
     get :new, survey_id: @survey, question_id: @question, answer_template_id: @answer_template
     assert_nil assigns(:survey)
@@ -48,7 +48,7 @@ class Builder::AnswerOptionsControllerTest < ActionController::TestCase
     assert_redirected_to root_path
   end
 
-  test "should create answer option as builder" do
+  test 'should create answer option as builder' do
     login(@builder)
     assert_difference('AnswerOption.count') do
       post :create, survey_id: @survey, question_id: @question, answer_template_id: @answer_template, answer_option: { text: 'New Answer Option', hotkey: '1', value: '1' }
@@ -63,7 +63,7 @@ class Builder::AnswerOptionsControllerTest < ActionController::TestCase
     assert_redirected_to builder_survey_question_answer_template_answer_option_path(assigns(:survey), assigns(:question), assigns(:answer_template), assigns(:answer_option))
   end
 
-  test "should not create answer option without text" do
+  test 'should not create answer option without text' do
     login(@builder)
     assert_difference('AnswerOption.count', 0) do
       post :create, survey_id: @survey, question_id: @question, answer_template_id: @answer_template, answer_option: { text: '', hotkey: '1', value: '1' }
@@ -78,7 +78,7 @@ class Builder::AnswerOptionsControllerTest < ActionController::TestCase
     assert_response :success
   end
 
-  test "should not create answer option as regular user" do
+  test 'should not create answer option as regular user' do
     login(@regular_user)
     assert_difference('AnswerOption.count', 0) do
       post :create, survey_id: @survey, question_id: @question, answer_template_id: @answer_template, answer_option: { text: 'New Answer Option', hotkey: '1', value: '1' }
@@ -90,7 +90,7 @@ class Builder::AnswerOptionsControllerTest < ActionController::TestCase
     assert_redirected_to root_path
   end
 
-  test "should show answer option as builder" do
+  test 'should show answer option as builder' do
     login(@builder)
     get :show, survey_id: @survey, question_id: @question, answer_template_id: @answer_template, id: @answer_option
     assert_not_nil assigns(:survey)
@@ -100,7 +100,7 @@ class Builder::AnswerOptionsControllerTest < ActionController::TestCase
     assert_response :success
   end
 
-  test "should not show answer option as regular user" do
+  test 'should not show answer option as regular user' do
     login(@regular_user)
     get :show, survey_id: @survey, question_id: @question, answer_template_id: @answer_template, id: @answer_option
     assert_nil assigns(:survey)
@@ -110,7 +110,7 @@ class Builder::AnswerOptionsControllerTest < ActionController::TestCase
     assert_redirected_to root_path
   end
 
-  test "should get edit answer option as builder" do
+  test 'should get edit answer option as builder' do
     login(@builder)
     get :edit, survey_id: @survey, question_id: @question, answer_template_id: @answer_template, id: @answer_option
     assert_not_nil assigns(:survey)
@@ -120,7 +120,7 @@ class Builder::AnswerOptionsControllerTest < ActionController::TestCase
     assert_response :success
   end
 
-  test "should not get edit answer option as regular user" do
+  test 'should not get edit answer option as regular user' do
     login(@regular_user)
     get :edit, survey_id: @survey, question_id: @question, answer_template_id: @answer_template, id: @answer_option
     assert_nil assigns(:survey)
@@ -130,7 +130,7 @@ class Builder::AnswerOptionsControllerTest < ActionController::TestCase
     assert_redirected_to root_path
   end
 
-  test "should update answer option as builder" do
+  test 'should update answer option as builder' do
     login(@builder)
     patch :update, survey_id: @survey, question_id: @question, answer_template_id: @answer_template, id: @answer_option, answer_option: { text: 'Updated Answer Option', hotkey: '2', value: '2' }
     assert_not_nil assigns(:survey)
@@ -143,7 +143,7 @@ class Builder::AnswerOptionsControllerTest < ActionController::TestCase
     assert_redirected_to builder_survey_question_answer_template_answer_option_path(assigns(:survey), assigns(:question), assigns(:answer_template), assigns(:answer_option))
   end
 
-  test "should not update answer option without name" do
+  test 'should not update answer option without name' do
     login(@builder)
     patch :update, survey_id: @survey, question_id: @question, answer_template_id: @answer_template, id: @answer_option, answer_option: { text: '', hotkey: '2', value: '2' }
     assert_not_nil assigns(:survey)
@@ -156,7 +156,7 @@ class Builder::AnswerOptionsControllerTest < ActionController::TestCase
     assert_response :success
   end
 
-  test "should not update answer option as regular user" do
+  test 'should not update answer option as regular user' do
     login(@regular_user)
     patch :update, survey_id: @survey, question_id: @question, answer_template_id: @answer_template, id: @answer_option, answer_option: { text: 'Updated Answer Option', hotkey: '2', value: '2' }
     assert_nil assigns(:survey)
@@ -166,7 +166,7 @@ class Builder::AnswerOptionsControllerTest < ActionController::TestCase
     assert_redirected_to root_path
   end
 
-  test "should destroy answer option as builder" do
+  test 'should destroy answer option as builder' do
     login(@builder)
     assert_difference('AnswerOption.current.count', -1) do
       delete :destroy, survey_id: @survey, question_id: @question, answer_template_id: @answer_template, id: @answer_option
@@ -178,7 +178,7 @@ class Builder::AnswerOptionsControllerTest < ActionController::TestCase
     assert_redirected_to builder_survey_question_answer_template_path(assigns(:survey), assigns(:question), assigns(:answer_template))
   end
 
-  test "should not destroy answer option as regular user" do
+  test 'should not destroy answer option as regular user' do
     login(@regular_user)
     assert_difference('AnswerOption.current.count', 0) do
       delete :destroy, survey_id: @survey, question_id: @question, answer_template_id: @answer_template, id: @answer_option
@@ -190,4 +190,9 @@ class Builder::AnswerOptionsControllerTest < ActionController::TestCase
     assert_redirected_to root_path
   end
 
+  test 'should save option order' do
+    login(@builder)
+    post :reorder, survey_id: @survey, question_id: @question, answer_template_id: @answer_template, answer_option_ids: [ActiveRecord::FixtureSet.identify(:web_answer_template_02), ActiveRecord::FixtureSet.identify(:web_answer_template_01)], format: 'js'
+    assert_response :success
+  end
 end