+++ test/controllers/builder/answer_templates_controller_test.rb
@@ -9,7 +9,7 @@ class Builder::AnswerTemplatesControllerTest < ActionController::TestCase
     @answer_template = answer_templates(:web_answer_template)
   end
 
-  test "should get answer templates as builder" do
+  test 'should get answer templates as builder' do
     login(@builder)
     get :index, survey_id: @survey, question_id: @question
     assert_not_nil assigns(:survey)
@@ -17,7 +17,7 @@ class Builder::AnswerTemplatesControllerTest < ActionController::TestCase
     assert_redirected_to builder_survey_question_path(assigns(:survey), assigns(:question))
   end
 
-  test "should not get answer templates as regular user" do
+  test 'should not get answer templates as regular user' do
     login(@regular_user)
     get :index, survey_id: @survey, question_id: @question
     assert_nil assigns(:survey)
@@ -25,7 +25,7 @@ class Builder::AnswerTemplatesControllerTest < ActionController::TestCase
     assert_redirected_to root_path
   end
 
-  test "should get new answer template as builder" do
+  test 'should get new answer template as builder' do
     login(@builder)
     get :new, survey_id: @survey, question_id: @question
     assert_not_nil assigns(:survey)
@@ -34,7 +34,7 @@ class Builder::AnswerTemplatesControllerTest < ActionController::TestCase
     assert_response :success
   end
 
-  test "should not get new answer template as regular user" do
+  test 'should not get new answer template as regular user' do
     login(@regular_user)
     get :new, survey_id: @survey, question_id: @question
     assert_nil assigns(:survey)
@@ -43,7 +43,7 @@ class Builder::AnswerTemplatesControllerTest < ActionController::TestCase
     assert_redirected_to root_path
   end
 
-  test "should create answer template as builder" do
+  test 'should create answer template as builder' do
     login(@builder)
     assert_difference('AnswerTemplate.count') do
       post :create, survey_id: @survey, question_id: @question, answer_template: { name: 'My New Answer Template', template_name: 'checkbox', parent_answer_option_value: @answer_template.parent_answer_option_value, text: @answer_template.text }
@@ -58,7 +58,7 @@ class Builder::AnswerTemplatesControllerTest < ActionController::TestCase
     assert_redirected_to builder_survey_question_answer_template_path(assigns(:survey), assigns(:question), assigns(:answer_template))
   end
 
-  test "should not create answer template without text" do
+  test 'should not create answer template without text' do
     login(@builder)
     assert_difference('AnswerTemplate.count', 0) do
       post :create, survey_id: @survey, question_id: @question, answer_template: { name: '', template_name: 'checkbox', parent_answer_option_value: @answer_template.parent_answer_option_value, text: @answer_template.text }
@@ -72,7 +72,7 @@ class Builder::AnswerTemplatesControllerTest < ActionController::TestCase
     assert_response :success
   end
 
-  test "should not create answer template as regular user" do
+  test 'should not create answer template as regular user' do
     login(@regular_user)
     assert_difference('AnswerTemplate.count', 0) do
       post :create, survey_id: @survey, question_id: @question, answer_template: { name: 'My New Answer Template', template_name: 'checkbox', parent_answer_option_value: @answer_template.parent_answer_option_value, text: @answer_template.text }
@@ -83,7 +83,7 @@ class Builder::AnswerTemplatesControllerTest < ActionController::TestCase
     assert_redirected_to root_path
   end
 
-  test "should show answer template as builder" do
+  test 'should show answer template as builder' do
     login(@builder)
     get :show, survey_id: @survey, question_id: @question, id: @answer_template
     assert_not_nil assigns(:survey)
@@ -92,7 +92,7 @@ class Builder::AnswerTemplatesControllerTest < ActionController::TestCase
     assert_response :success
   end
 
-  test "should not show answer template as regular user" do
+  test 'should not show answer template as regular user' do
     login(@regular_user)
     get :show, survey_id: @survey, question_id: @question, id: @answer_template
     assert_nil assigns(:survey)
@@ -101,7 +101,7 @@ class Builder::AnswerTemplatesControllerTest < ActionController::TestCase
     assert_redirected_to root_path
   end
 
-  test "should get edit answer template as builder" do
+  test 'should get edit answer template as builder' do
     login(@builder)
     get :edit, survey_id: @survey, question_id: @question, id: @answer_template
     assert_not_nil assigns(:survey)
@@ -110,7 +110,7 @@ class Builder::AnswerTemplatesControllerTest < ActionController::TestCase
     assert_response :success
   end
 
-  test "should not get edit answer template as regular user" do
+  test 'should not get edit answer template as regular user' do
     login(@regular_user)
     get :edit, survey_id: @survey, question_id: @question, id: @answer_template
     assert_nil assigns(:survey)
@@ -119,7 +119,7 @@ class Builder::AnswerTemplatesControllerTest < ActionController::TestCase
     assert_redirected_to root_path
   end
 
-  test "should update answer template as builder" do
+  test 'should update answer template as builder' do
     login(@builder)
     patch :update, survey_id: @survey, question_id: @question, id: @answer_template, answer_template: { name: 'Updated Answer Template', template_name: 'date', parent_answer_option_value: @answer_template.parent_answer_option_value, text: @answer_template.text }
     assert_not_nil assigns(:survey)
@@ -132,7 +132,7 @@ class Builder::AnswerTemplatesControllerTest < ActionController::TestCase
     assert_redirected_to builder_survey_question_answer_template_path(assigns(:survey), assigns(:question), assigns(:answer_template))
   end
 
-  test "should not update answer template without name" do
+  test 'should not update answer template without name' do
     login(@builder)
     patch :update, survey_id: @survey, question_id: @question, id: @answer_template, answer_template: { name: '', template_name: 'date', parent_answer_option_value: @answer_template.parent_answer_option_value, text: @answer_template.text }
     assert_not_nil assigns(:survey)
@@ -144,7 +144,7 @@ class Builder::AnswerTemplatesControllerTest < ActionController::TestCase
     assert_response :success
   end
 
-  test "should not update answer template as regular user" do
+  test 'should not update answer template as regular user' do
     login(@regular_user)
     patch :update, survey_id: @survey, question_id: @question, id: @answer_template, answer_template: { name: 'Updated Answer Template', template_name: 'date', parent_answer_option_value: @answer_template.parent_answer_option_value, text: @answer_template.text }
     assert_nil assigns(:survey)
@@ -153,7 +153,7 @@ class Builder::AnswerTemplatesControllerTest < ActionController::TestCase
     assert_redirected_to root_path
   end
 
-  test "should destroy answer template as builder" do
+  test 'should destroy answer template as builder' do
     login(@builder)
     assert_difference('AnswerTemplate.current.count', -1) do
       delete :destroy, survey_id: @survey, question_id: @question, id: @answer_template
@@ -164,7 +164,7 @@ class Builder::AnswerTemplatesControllerTest < ActionController::TestCase
     assert_redirected_to builder_survey_question_path(assigns(:survey), assigns(:question))
   end
 
-  test "should not destroy answer template as regular user" do
+  test 'should not destroy answer template as regular user' do
     login(@regular_user)
     assert_difference('AnswerTemplate.current.count', 0) do
       delete :destroy, survey_id: @survey, question_id: @question, id: @answer_template
@@ -175,4 +175,9 @@ class Builder::AnswerTemplatesControllerTest < ActionController::TestCase
     assert_redirected_to root_path
   end
 
+  test 'should save template order' do
+    login(@builder)
+    post :reorder, survey_id: @survey, question_id: @question, answer_template_ids: [ActiveRecord::FixtureSet.identify(:web_answer_template)], format: 'js'
+    assert_response :success
+  end
 end